Add files via upload
This dashboard investigates the sales of a company over one year, categorized by products, supervisors, and EMP.  The dashboard gives information regarding the overall performance over the year of the company. For example, performances for all supervisors are quite fairly distributed, with Supervisor 1 being the most efficient among the for. 

Techniquely, this dashboard was built through applying visual creations, basic and advanced filters in Power BI.