Add files via upload
In this dataset, I used Power Query Editor to edit data in Excel and connect the Country, Data, and Product Dataset.

Through DAX, I built new column to calculate the percentage of Grand in Total, and new measures to calculate the Profit and Sales in $.

For the visualization, I used clustered column chart, pie chart, map, and matrix. I also applied basic and advanced filters for the dataset to filter important information.